requirements.txt
requests
streamlit
PyPDF2
langchain
openai
python-dotenv
langchain_openai
pypdf
langchain_community
chromadb